pyproject.toml
[tool.poetry]
name = "sdss-hal"
version = "0.6.0b2"
description = "High-level observing tool for SDSS-V (replaces SOP)"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/hal"
repository = "https://github.com/sdss/hal"
documentation = "https://sdss-hal.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "hal", from = "src" }
]
include = ["src/hal/etc/*"]

[tool.poetry.scripts]
hal = "hal.__main__:hal"

[tool.poetry.dependencies]
python = "^3.9,<4.0"
sdsstools = "^1.0.0"
sdss-clu = "^1.4.0"
click-default-group = "^1.2.2"
numpy = "^1.22.1"
sdssdb = "^0.6.0"

[tool.poetry.dev-dependencies]
ipython = ">=8.0.0"
matplotlib = ">=3.1.1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = ">=3.3.1"
black = ">=20.8b1"
jedi = ">=0.17.2"
Sphinx = ">=4.0.0"
furo = ">=2021.9.22"
sphinx-copybutton = ">=0.4.0"
sphinx-click = ">=3.0.1"
sphinx-jsonschema = ">=1.16.11"
myst-parser = ">=0.15.2"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"

[tool.black]
line-length = 88
target-version = ['py310']
fast = true

[tool.isort]
profile = "black"
sections = ["FUTURE", "STDLIB", "TYPING", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "hal"
known_typing = ["typing"]
known_sdss = ["sdsstools", "clu"]
lines_after_imports = 2

[tool.pytest.ini_options]
addopts = "--cov hal --cov-report xml --cov-report html --cov-report term -W ignore"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
include = ["src/hal/*"]
omit = [
    "*/__init__.py",
    "src/hal/__main__.py",
    "src/hal/exceptions.py"
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
    "# pragma: no cover"
]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
